Move takeout handling to its own namespace
