+++ src/takeout/model.ts
@@ -1,0 +1,4 @@
+export interface ITakeoutRequest {
+    episodes: number;
+    seriesId: string;
+}
